[AIRLFOW-YYY] Move user and roles command to seperate files
